my_agent/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pyyaml==6.0.1
